Create main
The class QueryNotInCharacterRange() and function makeQuery() are optional. They are used to create a query that is 512 characters or less from a list of keywords in an excel sheet. If you would also like to import search terms from an excel sheet, some things to keep in mind:
A SPACE (" ") in a query is seen as and. 
OR is added between words here, but you can change that to a space to use and
see this link: https://developer.twitter.com/en/docs/twitter-api/tweets/search/integrate/build-a-query for more information on logic and restraints